Move `IndicatifFeedbackReceiver` to a separated file
