+++ mithril-client-cli/src/utils/mod.rs
@@ -1,10 +1,12 @@
 //! Utilities module
 //! This module contains tools needed mostly in services layers.
 
+mod feedback_receiver;
 mod progress_reporter;
 mod snapshot;
 mod unpacker;
 
+pub use feedback_receiver::*;
 pub use progress_reporter::*;
 pub use snapshot::*;
 pub use unpacker::*;
